Fix pattern parsing not extracting patterns correctly
If an invalid spawnable type was present in the XML file, then
the type would remain set to the type of the previous round,
continuing normally with the rest of the settings originally meant
for the invalid type. Now if an invalid type is detected, an
exception is thrown.

Also, if a spawnable type instead of containig multiple "<pattern>"
nodes it appears multiple times in the XML file, for every time the
type appears in the file another dimension is added to the "groupsByType"
collection; all patterns of the same type should be grouped in the
same dimension of the collection. To fix this, the "groupsByType"
collection was changed from a stack to a dictionary, so if a type
is repeated later we can append the collection by using
"groupsByType[type]".

Finally, the multiplier set using the "repeat" attribute in the XML
file was being set but never actually used.

Signed-off-by: AnimaRain <ioannis_spyr@yahoo.gr>